FEAT:06.07.24 V1

+++ src/main/java/com/cossinest/homes/service/business/CategoryService.java
@@ -10,12 +10,15 @@ import com.cossinest.homes.payload.mappers.CategoryMapper;
 import com.cossinest.homes.payload.messages.ErrorMessages;
 import com.cossinest.homes.exception.ConflictException;
 import com.cossinest.homes.exception.ResourceNotFoundException;
+import com.cossinest.homes.payload.messages.SuccesMessages;
 import com.cossinest.homes.payload.request.business.AdvertRequest;
 import com.cossinest.homes.payload.request.business.CategoryRequestDTO;
+import com.cossinest.homes.payload.response.ResponseMessage;
 import com.cossinest.homes.payload.response.business.AdvertResponse;
 import com.cossinest.homes.payload.response.business.CategoryPropKeyResponseDTO;
 import com.cossinest.homes.payload.response.business.CategoryPropKeyssResponseDTO;
 import com.cossinest.homes.payload.response.business.CategoryResponseDTO;
+import com.cossinest.homes.repository.business.AdvertRepository;
 import com.cossinest.homes.repository.business.CategoryPropertyKeyRepository;
 import com.cossinest.homes.repository.business.CategoryRepository;
 import jakarta.servlet.http.HttpServletRequest;
@@ -26,6 +29,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -35,7 +39,7 @@ import java.util.Objects;
 
 
 @RequiredArgsConstructor
-@Service
+@Service                            // @Service anot icinde @Component anot. da var
 public class CategoryService {
 
 
@@ -43,6 +47,7 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
     private final CategoryPropertyKeyRepository categoryPropertyKeyRepository;
+    private final AdvertRepository advertRepository;
 
 
 
@@ -66,8 +71,8 @@ public class CategoryService {
         Pageable pageable= PageRequest.of(page, size, Sort.by(String.valueOf(sort)).ascending());
 
         if(Objects.equals(type, "DESC")){
-                                            pageable=PageRequest.of(page, size, Sort.by(String.valueOf(sort)).descending());
-                                            }
+            pageable=PageRequest.of(page, size, Sort.by(String.valueOf(sort)).descending());
+        }
         return categoryRepository.findAllActiveCategories(pageable);
     }
 
@@ -82,38 +87,49 @@ public class CategoryService {
         return categoryRepository.findAllCategories(pageable);
     }
 
+
     public Category findCategoryById(Long id){
         return categoryRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("Category not found with id :" + id));
     }
 
 
-    public CategoryResponseDTO findCategoryWithId(Long id) {
+    public ResponseMessage<CategoryResponseDTO> findCategoryWithId(Long id) {
 
        Category category = categoryRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("Category not found with id :" + id));
-        return categoryMapper.mapCategoryToCategoryResponseDTO(category);
+
+        return ResponseMessage.<CategoryResponseDTO> builder()
+                .object(categoryMapper.mapCategoryToCategoryResponseDTO(category))
+                .message(SuccesMessages.RETURNED_A_CATEGORY)
+                .status(HttpStatus.OK)
+                .build();
     }
 
 
-    public CategoryResponseDTO createCategory(CategoryRequestDTO categoryRequestDTO) {
+    public ResponseMessage<CategoryResponseDTO> createCategory(CategoryRequestDTO categoryRequestDTO) {
 
         if(categoryRepository.existsByTitle(categoryRequestDTO.getTitle())) {
 
             throw new ConflictException("Category " +categoryRequestDTO.getTitle() +" is already exist " );
         }
+
         Category category = categoryMapper.mapCategoryRequestDTOToCategory(categoryRequestDTO);
 
         Category createdCategory = categoryRepository.save(category);
         createdCategory.generateSlug();
-        Category categoryresp = categoryRepository.save(createdCategory);
-        return categoryMapper.mapCategoryToCategoryResponseDTO(categoryresp);
+        Category categoryResponse = categoryRepository.save(createdCategory);
 
+        return ResponseMessage.<CategoryResponseDTO> builder()
+                .object(categoryMapper.mapCategoryToCategoryResponseDTO(categoryResponse))
+                .message(SuccesMessages.CATEGORY_CREATED_SUCCESS)
+                .status(HttpStatus.CREATED)
+                .build();
     }
 
 
 
-    public CategoryResponseDTO updateCategory(Long id, CategoryRequestDTO categoryRequestDTO) {
+    public ResponseMessage<CategoryResponseDTO> updateCategory(Long id, CategoryRequestDTO categoryRequestDTO) {
 
         boolean existTitle = categoryRepository.existsByTitle(categoryRequestDTO.getTitle());  // categoryRequest ile gelen Title DB'de VAR mi?
 
@@ -133,21 +149,21 @@ public class CategoryService {
         3.senaryo: categoryRequest ile gelen Title Daire ama DB de YOK,  DB'de MEVCUT Title : Arsa            --> FALSE && TRUE    (UPDATE OLUR)    */
         }
 
-        category.setTitle(categoryRequestDTO.getTitle());
-        category.setIcon(categoryRequestDTO.getIcon());
-        category.setSeq(categoryRequestDTO.getSeq());
-        category.setSlug(categoryRequestDTO.getSlug());
-        category.setActive(categoryRequestDTO.isActive());
+        // DTO ===>>> POJO
 
-        LocalDateTime updatedOn = LocalDateTime.now();
-        category.setUpdatedAt(updatedOn);
+        category = categoryMapper.mapCategoryRequestDTOToCategory(categoryRequestDTO);
 
         Category updatedCategory = categoryRepository.save(category);
-        return categoryMapper.mapCategoryToCategoryResponseDTO(updatedCategory);
+
+        return ResponseMessage.<CategoryResponseDTO> builder()
+                .object(categoryMapper.mapCategoryToCategoryResponseDTO(updatedCategory))
+                .message(SuccesMessages.CATEGORY_UPDATED_SUCCESS)
+                .status(HttpStatus.OK)
+                .build();
     }
 
 
-    public CategoryResponseDTO deleteCategory(Long id) {
+    public ResponseMessage<CategoryResponseDTO> deleteCategory(Long id) {
 
         Category category = categoryRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("Category not found with id :" + id));
@@ -156,25 +172,38 @@ public class CategoryService {
             throw new ResourceNotFoundException("This category cannot be deleted");
         }
 
+       Advert advert = advertRepository.findAdvertByCategory(id);
+
+        if (advert!=null){
+            throw new RuntimeException("This category cannot be deleted");
+        }
         categoryRepository.deleteById(id);
 
-        return categoryMapper.mapCategoryToCategoryResponseDTO(category);
+        return ResponseMessage.<CategoryResponseDTO> builder()
+                .object(categoryMapper.mapCategoryToCategoryResponseDTO(category))
+                .message(SuccesMessages.CATEGORY_DELETED_SUCCESS)
+                .status(HttpStatus.OK)
+                .build();
     }
 
 
-    public CategoryPropKeyssResponseDTO findCategoryPropertyKeys(Long id) {
+    public ResponseMessage<CategoryPropKeyssResponseDTO> findCategoryPropertyKeys(Long id) {
 
         Category category = findCategoryById(id);
 
         List<CategoryPropertyKey> categoryProperKeys = category.getCategoryPropertyKeys();
 
-        return categoryMapper.mapCategPropKeyssToCategPropKeyssResponseDTO(categoryProperKeys);
+        return ResponseMessage.<CategoryPropKeyssResponseDTO> builder()
+                .object(categoryMapper.mapCategPropKeyssToCategPropKeyssResponseDTO(categoryProperKeys))
+                .message(SuccesMessages.RETURNED_A_CATEGORY_PROPERTY_KEYS)
+                .status(HttpStatus.OK)
+                .build();
     }
 
 
 
 
-    public CategoryPropKeyResponseDTO createPropertyKey(Long id, String... keys) {
+    public  ResponseMessage<CategoryPropKeyResponseDTO> createPropertyKey(Long id, String... keys) {
 
 
         Category category = findCategoryById(id);
@@ -189,16 +218,24 @@ public class CategoryService {
         }
         categoryPropertyKeyRepository.save(categoryPropertyKey);
 
-        return categoryMapper.mapCategPropKeyToCategPropKeyResponseDTO(categoryPropertyKey);
+        return ResponseMessage.<CategoryPropKeyResponseDTO> builder()
+                .object(categoryMapper.mapCategPropKeyToCategPropKeyResponseDTO(categoryPropertyKey))
+                .message(SuccesMessages.CATEGORY_PROPERTY_KEY_CREATED_SUCCESS)
+                .status(HttpStatus.CREATED)
+                .build();
     }
 
 
-    public CategoryResponseDTO findCategoryBySlug(String slug) {
+    public ResponseMessage<CategoryResponseDTO> findCategoryBySlug(String slug) {
 
         Category category = categoryRepository.findBySlug(slug).orElseThrow(
                 ()-> new ResourceNotFoundException("Category not found with Slug :" + slug));
 
-        return categoryMapper.mapCategoryToCategoryResponseDTO(category);
+        return ResponseMessage.<CategoryResponseDTO> builder()
+                .object(categoryMapper.mapCategoryToCategoryResponseDTO(category))
+                .message(SuccesMessages.RETURNED_CATEGORY_BY_SLUG_SUCCESS)
+                .status(HttpStatus.OK)
+                .build();
     }
